+++ src/str/str.h
@@ -1,22 +1,22 @@
 #ifndef STR_H
 #define STR_H
 
-#include <stddef.h>
 #include <stdbool.h>
+#include <stdint.h>
 
 // Returns the number of bytes between `str` and the next `\0` character.
 //
 // Safety:
 // It is the caller's responsiblity to ensure that `str` is a valid string terminated by a
 // `\0` character. Failure to do so will result in invalid memory access.
-size_t ft_strlen(const char *str);
+uint64_t ft_strlen(const char *str);
 
 // Returns a heap-allocated duplicate of `str[start..start + len]`.
 //
 // Safety:
 // It is the caller's responsiblity to ensure that `str` is a valid string terminated by a
 // `\0` character. Failure to do so will result in invalid memory access.
-char *ft_substr(char const *str, unsigned int start, size_t len);
+char *ft_substr(char const *str, unsigned int start, uint64_t len);
 
 // Allocates and returns a new string with characters specified in `set` removed
 // from the beginning and end of the input string `str`
@@ -46,7 +46,7 @@ char **ft_split(char const *str, char delimiter);
 // Safety:
 // It is the caller's responsiblity to ensure that `needle` and `haystack` are valid strings terminated by
 // `\0` characters. Failure to do so will result in invalid memory access.
-char *ft_strnstr(const char *haystack, const char *needle, size_t n_bytes);
+char *ft_strnstr(const char *haystack, const char *needle, uint64_t n_bytes);
 
 // Converts all characters in `str` to lowercase.
 //
@@ -112,6 +112,6 @@ bool ft_atob(const char *nptr, int base, int bits, void *out_value, bool is_sign
 // Safety:
 // The caller must ensure that the `buffer` is large enough to hold the result. For example, a base 2
 // representation of a 64-bit integer requires up to 65 characters (64 digits + null terminator).
-bool ft_btoa(void *value, int base, int bits, char *buffer, size_t buffer_size, bool is_signed, bool uppercase);
+bool ft_btoa(void *value, int base, int bits, char *buffer, uint64_t buffer_size, bool is_signed, bool uppercase);
 
 #endif
